+++ backend/hardware/temperaturecontroller/basictempcontroller.py
@@ -1,11 +1,10 @@
-import config
-import hardware.thermometer as therm
 from hardware.temperaturecontroller.base import TempController
 import logging
 
 
 RELAY_ON = 1
 RELAY_OFF = 0
+
 
 class BasicTempController(TempController):
     def __init__(self, args, devices):
@@ -42,7 +41,7 @@ class BasicTempController(TempController):
         self.gpio.output(self.heaterPumpPin, RELAY_OFF)
         self.gpio.output(self.coolerPin, RELAY_OFF)
 
-        self.thermometer = devices[args['thermometerID']]
+        self.thermometer = devices[args["thermometerID"]]
 
     def turnHeaterOn(self):
         """
@@ -51,9 +50,8 @@ class BasicTempController(TempController):
         :return:
         None
         """
-        logging.info("heater turned on")
+        logging.debug("heater turned on")
         self.gpio.output(self.heaterPin, RELAY_ON)
-
 
     def turnHeaterOff(self):
         """
@@ -62,15 +60,15 @@ class BasicTempController(TempController):
         :return:
         None
         """
-        logging.info("heater turned off")
+        logging.debug("heater turned off")
         self.gpio.output(self.heaterPin, RELAY_OFF)
 
     def turnHeaterPumpOn(self):
-        logging.info("heater pump turned on")
+        logging.debug("heater pump turned on")
         self.gpio.output(self.heaterPumpPin, RELAY_ON)
 
     def turnHeaterPumpOff(self):
-        logging.info("heater pump turned off")
+        logging.debug("heater pump turned off")
         self.gpio.output(self.heaterPumpPin, RELAY_OFF)
 
     def turnCoolerOn(self):
@@ -80,9 +78,8 @@ class BasicTempController(TempController):
         :return:
         None
         """
-        logging.info("cooler turned on")
+        logging.debug("cooler turned on")
         self.gpio.output(self.coolerPin, RELAY_ON)
-
 
     def turnCoolerOff(self):
         """
@@ -91,9 +88,8 @@ class BasicTempController(TempController):
         :return:
         None
         """
-        logging.info("cooler turned off")
+        logging.debug("cooler turned off")
         self.gpio.output(self.coolerPin, RELAY_OFF)
-
 
     def getTemp(self):
         """
@@ -102,7 +98,7 @@ class BasicTempController(TempController):
         The temperature of the temperature sensor.
         """
         return self.thermometer.getTemp()
-        
+
     def getMaxTemperature(self):
         return self.maxTemp
 
@@ -111,4 +107,3 @@ class BasicTempController(TempController):
 
     def getPIDConfig(self):
         return self.pidConfig
-