+++ CHANGELOG.md
@@ -4,11 +4,12 @@
 
 ### Added
 
-- Many pre-release tweaks.
+- Many pre-release tweaks and bug fixes.
 - Massive docs update.
 - New printable cases, stands, and reactor parts.
 - Ability to manually reload hardware from the settings menu. (#197)
 - Hardware settings are configurable from the UI in a basic fashion (#141).
+- v0.6 hardware config supports PID temperature control
 - MicroLab 0.5 simulation configuration (#140).
 - Configurable hardware limits for supported temperatures. Recipes that require values outside of the configured range will no longer run. (#146).
 - Sounds that play on recipe completion, errors, and user input steps (#149).
@@ -18,6 +19,7 @@
 
 ### Changed
 
+- Performance and logging tuning
 - Configuration is now stored in /etc/microlab instead of backend/config.py (#141).
 - Recipe data was moved from /backend/recipes/files to being stored by default at /var/lib/microlab/recipes. Recipes in the repository are kept at backend/data/recipes and copied to their proper storage location at launch. (#141).
 - Hardware configuration files were moved from backend/hardware to being stored by default at /var/lib/microlab/controllerhardware and /var/lib/microlab/labhardware. Hardware configurations in the repository are kept at backend/data/hardware and copied to their proper storage location at launch. (#141).